Simplify.
