+++ heron/common/tests/java/BUILD
@@ -20,6 +20,7 @@ java_tests(
         "com.twitter.heron.common.basics.SysUtilsTest",
         "com.twitter.heron.common.basics.WakeableLooperTest",
         "com.twitter.heron.common.basics.FileUtilsTest",
+        "com.twitter.heron.common.basics.PackageTypeTest",
         "com.twitter.heron.common.network.EchoTest",
         "com.twitter.heron.common.network.HeronServerTest",
         "com.twitter.heron.common.config.ConfigReaderTest",
