+++ heron/common/tests/java/com/twitter/heron/common/basics/FileUtilsTest.java
@@ -121,17 +121,7 @@ public class FileUtilsTest {
     Assert.assertFalse(FileUtils.isDirectoryExists(""));
   }
 
-  /**
-   * Method: isOriginalPackageJar(String packageFilename)
-   */
-  @Test
-  public void testIsOriginalPackageJar() throws Exception {
-    String jarFile = "a.jar";
-    Assert.assertTrue(FileUtils.isOriginalPackageJar(jarFile));
 
-    String notJarFile = "b.tar";
-    Assert.assertFalse(FileUtils.isOriginalPackageJar(notJarFile));
-  }
 
   /**
    * Method: getBaseName(String file)
